fix type latitude longitude from Int to Double

+++ lib/src/main/java/com/trueelogistics/checkin/model/HubLocationPointModel.kt
@@ -2,5 +2,5 @@ package com.trueelogistics.checkin.model
 
 data class HubLocationPointModel(
     var type : String? = null ,
-    var coordinates : ArrayList<Int>
+    var coordinates : ArrayList<Double>
 )
